[WIP] Cleanup - compartmentalized some functions and improved collections API in webgl.js
