+++ package.json
@@ -22,7 +22,6 @@
   "dependencies": {
     "@barba/core": "^2.9.7",
     "comlink": "^4.3.1",
-    "imagesloaded": "^5.0.0",
     "motion": "^10.5.2",
     "nanostores": "^0.7.0",
     "normalize-wheel": "^1.0.1",
