+++ src/lib/lerp.js
@@ -1,0 +1,3 @@
+export function lerp(p1, p2, t) {
+  return p1 + (p2 - p1) * t;
+}
